Flipping-the-matrix-statement
Problem Statement:
Sean invented a game involving a 2nx2n matrix where each cell of the matrix contains an integer. He can reverse any of its rows or columns any number of times. The goal of the game is to maximize the sum of the elements in the nxn submatrix located in the upper-left quadrant of the matrix.

Given the initial configurations for matrices, help Sean reverse the rows and columns of each matrix in the best possible way so that the sum of the elements in the matrix's upper-left quadrant is maximal. 

For more details refer:
https://www.hackerrank.com/contests/target-codevita-11/challenges/flipping-the-matrix