Staircase-Statement
Problem Statement:
Consider a staircase of size n=4:
      #
    ##
  ###
####

Observe that its base and height are both equal to n, and the image is drawn using # symbols and spaces. The last line is not preceded by any spaces.
Write a program that prints a staircase of size n.

Function Description:
Complete the staircase function in the editor below. It should print a staircase as described above.
staircase has the following parameter(s):
    n: an integer

For more details refer
https://www.hackerrank.com/challenges/staircase/problem